Refactor: kueue module
