+++ src/codeflare_sdk/ray/cluster/generate_yaml.py
@@ -18,18 +18,17 @@ This sub-module exists primarily to be used internally by the Cluster object
 """
 
 import json
-from typing import Optional
 import typing
 import yaml
 import os
 import uuid
 from kubernetes import client
 from ...common import _kube_api_error_handling
+from ...common.kueue import add_queue_label
 from ...common.kubernetes_cluster.auth import (
     get_api_client,
     config_check,
 )
-from kubernetes.client.exceptions import ApiException
 import codeflare_sdk
 
 
@@ -227,65 +226,6 @@ def update_nodes(
 
 def del_from_list_by_name(l: list, target: typing.List[str]) -> list:
     return [x for x in l if x["name"] not in target]
-
-
-def get_default_kueue_name(namespace: str):
-    # If the local queue is set, use it. Otherwise, try to use the default queue.
-    try:
-        config_check()
-        api_instance = client.CustomObjectsApi(get_api_client())
-        local_queues = api_instance.list_namespaced_custom_object(
-            group="kueue.x-k8s.io",
-            version="v1beta1",
-            namespace=namespace,
-            plural="localqueues",
-        )
-    except ApiException as e:  # pragma: no cover
-        if e.status == 404 or e.status == 403:
-            return
-        else:
-            return _kube_api_error_handling(e)
-    for lq in local_queues["items"]:
-        if (
-            "annotations" in lq["metadata"]
-            and "kueue.x-k8s.io/default-queue" in lq["metadata"]["annotations"]
-            and lq["metadata"]["annotations"]["kueue.x-k8s.io/default-queue"].lower()
-            == "true"
-        ):
-            return lq["metadata"]["name"]
-
-
-def local_queue_exists(namespace: str, local_queue_name: str):
-    # get all local queues in the namespace
-    try:
-        config_check()
-        api_instance = client.CustomObjectsApi(get_api_client())
-        local_queues = api_instance.list_namespaced_custom_object(
-            group="kueue.x-k8s.io",
-            version="v1beta1",
-            namespace=namespace,
-            plural="localqueues",
-        )
-    except Exception as e:  # pragma: no cover
-        return _kube_api_error_handling(e)
-    # check if local queue with the name provided in cluster config exists
-    for lq in local_queues["items"]:
-        if lq["metadata"]["name"] == local_queue_name:
-            return True
-    return False
-
-
-def add_queue_label(item: dict, namespace: str, local_queue: Optional[str]):
-    lq_name = local_queue or get_default_kueue_name(namespace)
-    if lq_name == None:
-        return
-    elif not local_queue_exists(namespace, lq_name):
-        raise ValueError(
-            "local_queue provided does not exist or is not in this namespace. Please provide the correct local_queue name in Cluster Configuration"
-        )
-    if not "labels" in item["metadata"]:
-        item["metadata"]["labels"] = {}
-    item["metadata"]["labels"].update({"kueue.x-k8s.io/queue-name": lq_name})
 
 
 def augment_labels(item: dict, labels: dict):
